Add experimental EddystoneService
Add a new implementation for EddystoneService that will replace both
EddystoneConfigService.h and EddystoneService.h in ble/services. The new
implementation is built from scratch and is expected to reduce memory
usage and be more configurable. Furthermore, when used with mbed OS it will
use minar callbacks rather than Ticker or Timeouts to avoid race conditions.