SM example: defer setLinkSecurity.
There's an issue with current mbed-os codebase where the security
manager is notified of the existence of a new connection after the
user event handler. This prevent correct execution of setLinkSecurity
in the connection event handler.

A simple workaround i to defer execution of setLinkSecurity after
the execution the user event handler.