Add onPrivacyEnabled handler and explanatory comment
If "app.filesystem-support" is configured to true, privacy is then enabled through the SecurityManager API. Before this commit, an unexplained delay is used to wait until the privacy system is ready for GAP operations in this case.

This commit introduces an explanation as to why there is a deferred call to start advertising and exemplifies the use of Gap::EventHandler::onPrivacyEnabled.

Resolves #335.